Move useSelectedItemsPushSize into modules sync hooks folder
